split the stats out from gifts and added the event stats output
